refactor(@crates/eur-timeline): refactor the x11 focus tracker to make room for multiple impl's
