+++ crates/app/eur-timeline/src/linux/mod.rs
@@ -1,0 +1,4 @@
+mod utils;
+mod xorg_focus_tracker;
+
+pub mod impl_focus_tracker;
