+++ crates/app/eur-timeline/src/linux/utils.rs
@@ -1,0 +1,15 @@
+use std::env::var_os;
+
+pub fn wayland_detect() -> bool {
+    let xdg_session_type = var_os("XDG_SESSION_TYPE")
+        .unwrap_or_default()
+        .to_string_lossy()
+        .to_string();
+
+    let wayland_display = var_os("WAYLAND_DISPLAY")
+        .unwrap_or_default()
+        .to_string_lossy()
+        .to_string();
+
+    xdg_session_type.eq("wayland") || wayland_display.to_lowercase().contains("wayland")
+}
